# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "U1R API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Get Order List",
                "operationId": "52e2e0056ce273258998e1545a77a151",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "user id",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "order_status",
                        "in": "query",
                        "description": "Order status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Pending",
                                "Confirmed"
                            ]
                        }
                    },
                    {
                        "name": "payment_status",
                        "in": "query",
                        "description": "Payment status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Unpaid",
                                "Paid",
                                "PartialPaid"
                            ]
                        }
                    },
                    {
                        "name": "shipping_status",
                        "in": "query",
                        "description": "Shipping status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Processing",
                                "OnTheWay",
                                "Delivered",
                                "Returned"
                            ]
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Get Order List successfully"
                    }
                }
            },
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Submit Order",
                "operationId": "69a48b8c7cb177955d5c615348201c4e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "phone_no",
                                    "city",
                                    "shipping_address",
                                    "payment_id"
                                ],
                                "properties": {
                                    "first_name": {},
                                    "last_name": {},
                                    "email": {},
                                    "phone_no": {},
                                    "district": {},
                                    "city": {},
                                    "postal_code": {},
                                    "shipping_address": {},
                                    "payment_id": {
                                        "type": "number"
                                    },
                                    "stores": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "store_id": {
                                                    "type": "integer"
                                                },
                                                "shipping_cost": {
                                                    "type": "integer"
                                                },
                                                "discount_type": {},
                                                "discount": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "product_id": {
                                                                "type": "integer"
                                                            },
                                                            "product_variant_id": {
                                                                "type": "integer"
                                                            },
                                                            "quantity": {
                                                                "type": "integer"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order submitted Successfully"
                    }
                }
            }
        },
...
        "/api/orders/{id}": {
            "patch": {
                "tags": [
                    "Order"
                ],
                "summary": "Update Order Status",
                "operationId": "2102145d2d36d19d5fc27c247c3dd421",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "payment_status": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated Successfully"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}